pyproject.toml
[project]
name = "rca-agent"
version = "0.1.0"
description = "Intelligent Root Cause Analysis Agent for Data Engineering Pipelines"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Mohamed Hedi FERJANI", email = "ferjani.mohamedhedi@gmail.com" }
]
keywords = [
    "data-engineering",
    "rca",
    "root-cause-analysis",
    "airflow",
    "llm",
    "agent",
    "mlops",
    "observability"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "langgraph>=0.2.62",
    "langchain>=0.3.14",
    "langchain-anthropic>=0.3.1",
    "langchain-openai>=0.3.0",
    "chromadb>=0.5.23",
    "httpx>=0.28.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "python-dotenv>=1.0.1",
    "gitpython>=3.1.44",
    "jinja2>=3.1.5",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
huggingface = [
    "langchain-huggingface>=0.1.2",
    "huggingface-hub>=0.27.1",
    "transformers>=4.47.1",
    "accelerate>=1.2.1",
    "langchain-mistralai>=0.2.4",
]
ollama = [
    "langchain-ollama>=0.2.2",
]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.2",
    "pytest-mock>=3.14.0",
    "mypy>=1.14.1",
    "ruff>=0.8.6",
    "pre-commit>=4.0.1",
    "respx>=0.22.0",
]
docs = [
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-autodoc-typehints>=2.5.0",
    "myst-parser>=4.0.0",
]
airflow = [
    "apache-airflow>=2.10.4",
]
all = [
    "rca-agent[dev,docs,airflow,huggingface,ollama]",
]

[project.scripts]
rca-agent = "rca_agent.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/rca-agent"
Documentation = "https://rca-agent.readthedocs.io"
Repository = "https://github.com/yourusername/rca-agent"
Issues = "https://github.com/yourusername/rca-agent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rca_agent"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.2",
    "pytest-mock>=3.14.0",
    "mypy>=1.14.1",
    "ruff>=0.8.6",
    "pre-commit>=4.0.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["rca_agent"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "langgraph.*",
    "langchain.*",
    "langchain_anthropic.*",
    "langchain_openai.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--cov=src/rca_agent",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src/rca_agent"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
